requirements.txt
streamlit
requests
pydub
imageio-ffmpeg
google-genai
